feat: Add icon to main window and About menu item
This commit adds an icon to the main window of the application and includes an "About" menu item in the toolbar. The icon enhances the visual appeal of the application, while the "About" menu item provides users with information about the application. These changes improve the user experience and add important functionality to the application.